+++ app/src/main/java/com/dede/android_eggs/inject/EasterEggModules.kt
@@ -84,7 +84,7 @@ object EasterEggModules {
                     add(easterEgg)
                 }
             }
-        }
+        }.sortedByDescending { it.fullApiLevel }
     }
 
 }